refactor(authorship): revert prov tooltip styles
