+++ web/src/ui/nodes/properties/authorship/index.ts
@@ -1,0 +1,4 @@
+import './authorship'
+import { getTooltipContent } from './utils'
+
+export { getTooltipContent }
